requirements.txt
asgiref==3.2.3
bcrypt==3.1.7
cffi==1.14.0
Django==3.0.3
django-contrib-comments==1.9.2
django-js-asset==1.2.2
django-mptt==0.11.0
django-widget-tweaks==1.4.5
pycparser==2.19
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
